# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Tadweerah API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /companies:
    post:
      operationId: createCompany
      tags: [companies]
      summary: Create a company for the current user (one company per user in MVP)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCompanyBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
        "400":
          description: Validation error
        "401":
          description: Unauthorized
        "409":
          description: User already has a company
  /companies/mine/capabilities:
    get:
      operationId: getMyCapabilities
      tags: [companies]
      summary: Get the current company's declared capabilities.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CompanyCapabilityRow"
        "401":
          description: Unauthorized
        "403":
          description: Company required
    put:
      operationId: updateMyCapabilities
      tags: [companies]
      summary: Replace the current company's capability list.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCapabilitiesBody"
      responses:
        "200":
          description: Updated list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CompanyCapabilityRow"
        "400":
          description: Validation error
        "401":
          description: Unauthorized
        "403":
          description: Company required
...
  /deals/{deal_id}:
    get:
      operationId: getDeal
      tags: [deals]
      summary: Get deal details. Accessible by producer or buyer of the deal.
      parameters:
        - in: path
          name: deal_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deal"
        "401":
          description: Unauthorized
        "403":
          description: Not a party to this deal
        "404":
          description: Deal not found
  /deals/{deal_id}/confirm-payment:
    post:
      operationId: confirmPayment
      tags: [deals]
      summary: >
        Producer confirms receipt of payment.
        Requires payment_reference (bank transfer / transaction ID).
        For by_weight deals, actual_quantity is required to compute final_amount.
        Transitions deal status from active → payment_confirmed.
      parameters:
        - in: path
          name: deal_id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfirmPaymentBody"
      responses:
        "200":
          description: Payment confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deal"
        "400":
          description: Missing payment_reference or invalid actual_quantity
        "401":
          description: Unauthorized
        "403":
          description: Not the producer or deal not in active state
        "404":
          description: Deal not found
  /deals/{deal_id}/confirm-dispatch:
    post:
      operationId: confirmDispatch
      tags: [deals]
      summary: >
        Producer confirms goods have been dispatched to the buyer.
        Transitions deal status from payment_confirmed → dispatched.
      parameters:
        - in: path
          name: deal_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Dispatch confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deal"
        "401":
          description: Unauthorized
        "403":
          description: Not the producer or deal not in payment_confirmed state
        "404":
          description: Deal not found
  /deals/{deal_id}/confirm-receipt:
    post:
      operationId: confirmReceipt
      tags: [deals]
      summary: >
        Buyer confirms receipt of goods.
        Transitions deal status from dispatched → completed.
      parameters:
        - in: path
          name: deal_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Receipt confirmed, deal completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deal"
        "401":
          description: Unauthorized
        "403":
          description: Not the buyer or deal not in dispatched state
        "404":
          description: Deal not found

  # ── Notifications ────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    CompanyType:
      type: string
      enum: [producer, buyer, carrier]
      description: producer = waste producer, buyer = recycling company, carrier = transporter
    LicenseStatus:
      type: string
      enum: [pending, approved, rejected, expired]
      description: |
        null = no license submitted (participates freely).
        pending = license submitted, awaiting admin review (no restrictions yet).
        approved = license verified.
        rejected / expired = license not valid.
    Company:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          $ref: "#/components/schemas/CompanyType"
        city:
          type: string
        commercialRegistration:
          type: string
        contactPhone:
          type: string
        license_number:
          type: string
          description: Regulatory license number (e.g. MOMRA, NCBE). Optional.
        license_status:
          $ref: "#/components/schemas/LicenseStatus"
        company_category_id:
          type: string
          description: FK to company_categories.id. Descriptive business category.
        createdAt:
          type: string
          format: date-time
      required: [id, name, type, city, contactPhone, createdAt]
    CreateCompanyBody:
      type: object
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 120
        type:
          $ref: "#/components/schemas/CompanyType"
        city:
          type: string
          minLength: 2
          maxLength: 80
        commercialRegistration:
          type: string
          maxLength: 40
        contactPhone:
          type: string
          minLength: 6
          maxLength: 20
        license_number:
          type: string
          maxLength: 80
          description: Optional regulatory license number.
        company_category_id:
          type: string
          description: Optional FK to company_categories.id.
        accepted_terms:
          type: boolean
          description: Must be true to accept platform terms and conditions.
      required: [name, type, city, contactPhone]
...
    CompanyCapabilityRow:
      type: object
      description: A capability entry as returned when reading a company's declared capabilities.
      properties:
        capability_id:
          type: string
        key:
          type: string
          description: Stable internal key (e.g. "recycle_metal"). Never changes.
        name_ar:
          type: string
        name_en:
          type: string
        sort_order:
          type: integer
      required: [capability_id, key, name_ar, name_en, sort_order]
    UpdateCapabilitiesBody:
      type: object
      properties:
        capability_ids:
          type: array
          items:
            type: string
          description: Full replacement list of capability UUIDs for this company.
      required: [capability_ids]

    # ── Deals ────────────────────────────────────────────────────────────────
    DealStatus:
      type: string
      enum: [active, payment_confirmed, dispatched, completed]
    DealSettlementType:
      type: string
      enum: [fixed, by_weight, revenue_share]
    DealCounterparty:
      type: object
      properties:
        name:
          type: string
        contact_phone:
          type: string
      required: [name, contact_phone]
    Deal:
      type: object
      description: >
        Represents a confirmed agreement between producer and buyer following offer acceptance.
        Progresses through: active → payment_confirmed → dispatched → completed.
      properties:
        id:
          type: string
        offer_id:
          type: string
        listing_id:
          type: string
        producer_company_id:
          type: string
        buyer_company_id:
          type: string
        settlement_type:
          $ref: "#/components/schemas/DealSettlementType"
        price_per_unit:
          type: number
        estimated_amount:
          type: number
          description: price_per_unit × listing quantity. Indicative only for by_weight deals.
        actual_quantity:
          type: number
          description: Filled by producer on payment confirmation for by_weight deals.
        final_amount:
          type: number
          description: price_per_unit × actual_quantity. Only set for by_weight deals after confirmation.
        status:
          $ref: "#/components/schemas/DealStatus"
        counterparty:
          $ref: "#/components/schemas/DealCounterparty"
          description: Contact info of the other party (producer sees buyer info and vice versa).
        payment_confirmed_at:
          type: string
          format: date-time
        payment_reference:
          type: string
          description: Bank transfer number / transaction reference provided by producer.
        payment_proof_url:
          type: string
          description: Optional URL to payment proof document/screenshot.
        dispatched_at:
          type: string
          format: date-time
        received_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        [id, offer_id, listing_id, producer_company_id, buyer_company_id,
         settlement_type, price_per_unit, estimated_amount, status, created_at, updated_at]
    ConfirmPaymentBody:
      type: object
      description: >
        Body for POST /deals/{deal_id}/confirm-payment.
        payment_reference is always required.
        actual_quantity is required for by_weight deals.
      properties:
        payment_reference:
          type: string
          minLength: 1
          description: Bank transfer number or transaction reference ID.
        payment_proof_url:
          type: string
          description: Optional URL to a payment screenshot or document.
        actual_quantity:
          type: number
          exclusiveMinimum: 0
          description: Actual weighed quantity. Required for by_weight deals.
      required: [payment_reference]

    # ── Admin write schemas ───────────────────────────────────────────────────
...