# lib/api-zod/src/generated/types/dealSettlementType.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tadweerah API
 * OpenAPI spec version: 0.1.0
 */

export type DealSettlementType =
  (typeof DealSettlementType)[keyof typeof DealSettlementType];

export const DealSettlementType = {
  fixed: "fixed",
  by_weight: "by_weight",
  revenue_share: "revenue_share",
} as const;